+++ core-ui/package.json
@@ -11,6 +11,10 @@
     "@peculiar/x509": "^1.6.0",
     "@sentry/react": "^6.17.4",
     "@stoplight/json-ref-resolver": "^3.1.3",
+    "@types/jest": "^29.1.2",
+    "@types/node": "^18.8.4",
+    "@types/react": "^18.0.21",
+    "@types/react-dom": "^18.0.6",
     "@ui-schema/ui-schema": "^0.4.0-beta.1",
     "buffer": "^6.0.3",
     "bytes-iec": "^3.1.0",
@@ -38,6 +42,7 @@
     "react-tippy": "^1.4.0",
     "recoil": "^0.7.5",
     "rfc6902": "^4.0.1",
+    "typescript": "^4.8.4",
     "use-debounce": "^7.0.1",
     "yup": "^0.28.5"
   },
