--- core-ui/src/state/openapiSchemasAtom.js
@@ -1,13 +1,0 @@
-import { atom } from 'recoil';
-
-const defaultValue = {
-  schemaInfo: {
-    areSchemasComputed: false,
-    schemasError: null,
-  },
-};
-
-export const openapiSchemasState = atom({
-  key: 'openapiSchemasState',
-  default: defaultValue,
-});
